+++ src/main/java/blackjack/domain/person/Person.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.person;
 
+import blackjack.domain.card.Card;
 import java.util.ArrayList;
 import java.util.List;
